+++ src/main/java/lotto/utils/RandomUtils.java
@@ -3,11 +3,10 @@ package lotto.utils;
 import java.util.Random;
 
 public class RandomUtils {
-
     public static final int LOTTERY_MAX_NUMBER = 45;
     public static final int LOTTERY_MIN_NUMBER = 1;
 
-    public static int random(){
-       return  (new Random().nextInt(LOTTERY_MAX_NUMBER)) + LOTTERY_MIN_NUMBER;
+    public static int random() {
+        return (new Random().nextInt(LOTTERY_MAX_NUMBER)) + LOTTERY_MIN_NUMBER;
     }
 }
