+++ src/main/java/lotto/vo/LotteryMaker.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.vo;
 
 import lotto.utils.RandomUtils;
